[noetic*] Remove deadsnakes ppa
The deadsnakes team no longer publishes packages for Ubuntu 20.04
(focal). The noetic Dockerfile dependencies on the deadsnakes ppa would
cause their builds to fail. An old version of Python3.9 is available in
the Ubuntu main repository -- Poetry requires 3.9+ as of this commit.

uv provides a fairly easy way to manage and run newer versions of
Python, and is recommended over Poetry at this time.